[Kernel] Fix int32 overflow in moe_fused_mul_sum kernel
Promote offs_m to int64 before the `offs_m * stride_m` workspace offset multiply
so the (num_tokens, top_k, hidden) expert-output workspace is addressed
correctly past 2**31 elements.

+++ vllm/model_executor/layers/fused_moe/moe_fused_mul_sum.py
@@ -32,7 +32,7 @@ def moe_fused_mul_sum_kernel(
     k_mask = offs_k < size
     mask = m_mask[:, None] & k_mask[None, :]
 
-    a_base = inputs_ptr + (offs_m * stride_m)[:, None] + offs_k[None, :]
+    a_base = inputs_ptr + (offs_m.to(tl.int64) * stride_m)[:, None] + offs_k[None, :]
     b_base = topk_weights_ptr + offs_m * top_k
 
     acc = tl.zeros((BLOCK_M, BLOCK_K), dtype=tl.float32)
